requirements.txt
addict==2.2.1
geopandas==0.8.2
pandas==1.0.5
PyYAML==5.4.1
rasterio==1.2.0
torch==1.7.1
torchvision==0.8.2
tensorboard==2.4.1
six==1.12.0
matplotlib==3.3.4